+++ src/qcommon/q_shared.h
@@ -1259,7 +1259,9 @@ typedef enum
 {
   CV_NONE = 0,
   CV_FLOAT,
-  CV_INTEGER
+  CV_INTEGER,
+  CV_FSPATH,
+  CV_MAX,
 }
 cvarValidator_t;