+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Changes
 -----------------------
 **--01/06/2026--**
 
+* qcommon:<br />added `CV_FSPATH` and `CV_MAX`
+
 * qcommon/server:<br />group cvars for better tracking
 
 * qcommon:<br />buffer and sock fixes to netchan